+++ CHANGELOG.md
@@ -10,7 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### New Features
 - DisplayTriangleShell3D: a TriangleShell3D optimized for performance of display / saving / loading.
 - nurbs module.
-- 
+- New curves classes: Hyperbola2D and Hyperbola3D.
+- Line: closest_point_on_line, from_point_and_vector
+- Line2D: get_slope, get_y_intersection.
+- New curves classes: Parabola2D/3D.
+- ConicalSurface3D: line/line_segment intersections, perpendicular_plane_intersection
 ### Fixed
 - add missing name attributes to classmethods
 - fixed circular imports
@@ -21,9 +25,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Cache BSplineCurve points into a numpy array to reduce memory usage.
+- 
 ### Unittests
--
-
+- Hyperbola2D/3D: line_intersections
+- Parabola2D/3D: line_intersections
+- ConicalSurface3D: test_line_intersections
 ## v0.13.0 [unreleased]
 
 ### New Features